Added album download option

+++ config/config.go
@@ -17,11 +17,11 @@ var Config struct {
 	// Your telegram bot token
 	BotToken string `json:"bot_token"`
 	// Authorized users to use this bot. Use @myidbot to get your ID
-	Users []int `json:"users"`
+	Users []int64 `json:"users"`
 }
 
 // Private map to check authorized users
-var users map[int]struct{}
+var users map[int64]struct{}
 
 // LoadConfig loads the config file from a location
 func LoadConfig(location string) {
@@ -34,7 +34,7 @@ func LoadConfig(location string) {
 		log.Fatalf("Cannot parse config file: %s\n", err)
 	}
 	// Populate users map
-	users = make(map[int]struct{}, len(Config.Users))
+	users = make(map[int64]struct{}, len(Config.Users))
 	for _, user := range Config.Users {
 		users[user] = struct{}{}
 	}
@@ -42,7 +42,7 @@ func LoadConfig(location string) {
 }
 
 // CheckAuthorizedUser checks if userId is available in users map and is allowed to use the bot
-func CheckAuthorizedUser(userId int) bool {
+func CheckAuthorizedUser(userId int64) bool {
 	_, exists := users[userId]
 	return exists
 }
